python/requirements.txt
web3==5.12.0
load-dotenv==0.1.0